requirements.txt
python-frontmatter==0.1.4
PyYAML==3.11
six==1.8.0
